uiuctf2020/Bot Protection IV/requirements.txt
beautifulsoup4==4.9.1
matplotlib==3.3.0
numpy==1.18.5
Pillow==7.1.2
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.2
torch==1.5.0
torchvision==0.6.0